requirements.txt
asgiref==3.9.1
django==4.0.2
django-markdownx==4.0.9
importlib-metadata==8.7.0
markdown==3.8.2
pillow==11.3.0
python-dotenv==1.1.1
sqlparse==0.5.3
typing-extensions==4.14.1
zipp==3.23.0
